create memoryutils class
